refactor: reorg all test files in the virtual-curve package
